+++ CHANGELOG.md
@@ -91,6 +91,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Better organization/separation for different steps in the window.
 - Improved settings tabs layout.
 - All image host uploaders now log retries.
+- **Overview** page initialization is now handled in a threaded worker, to keep the UI smooth while it's handling longer loading NFOs/plugins.
 
 ### Fixed